UI major fix Pt.1
Update:
fixed some of the UI resolution,
Fixed windowed options,
Fixed more forward.

Need to fix:
the rest of the UI folders need to be edited and some need to be elements need to be partied.

Need to rename the Control buttons and the keybind titles from more to move.