Mod: Planting Mechanics
Added in prototype of planting mechanic, including checking if the spot is occupied, and whether or not the player has the seeds for it, and if so, it will place it. For now, it is all communicated in the console until we can get the crops and other features going. All the code is mostly commented through and explained.