Mod: Time Mechanic Complete
Fully completed the time mechanic, including the time going by 1 minute per second, 1 hour per 60 minutes, 1 day per 24 hours, 1 week per 7 days, 1 month per 4 weeks, and 1 year per 12 months. Additionally, the light slowly rotates based on the total minutes of the game. As of right now, there is not functionality of the season, as that will directly link to the growing script, so i will merge to the dev branch, and then, once the growing script is done, I will merge it to the dev branch as well, then do some final touches to ensure that both mechanics are fully functional and working together.