Fix: Prefabs for Planting, growing and harvesting
Went through and deleted the prefabs that were there for planting, growing and harvesting, but will not be needed for the planting, growing and harvesting mechanics. I then also imported the new not prepped, dry and damp slot models as prefabs, then put them all into one prefab and attached the plot handler script to it. I then added colliders to the plot slots, and then put them on the farm plots and made that a prefab, then created 4 and put them in the scene