Mod: Made Growing Mechanic and some tweaks to watering and interaction
Made it so that the code fins out what type of crop it is, then fills out an array with an amount in which would be suitable for the days in which the crop would progress a stage. Then, when the day reaches the number, it will progress a stage. Still need to do some testing to make sure it is all working, and properly implement planting.