Mod: Finished PGH, but need to implement inventory with it
I have completed the Plant Grow Harvest mechanics, including being able to make it so that when the player tries to plant a seed, it will check if the plot is in not in a not prepped state and that there is enough of that seed. If so, it will plant the seed, changing the current crop that is on the plot to that specific crop, and changing the state of the crop to planted. This mechanic is attached to the plot slot object.

Then, for growing, we have a basic system where a switch system goes through to determine what set days the plant will continue through a stage, and then it will go through and check if the day has been reached and switch to a specific stage if so.

For the harvest, we have it so that the adult crop has a collider attached to it, and when interacting with it, the current crop of the plot will be set to none, and the crop will be added to inventory.

However, some main things are that I was struggling to figure out the inventory mechanic, so i am going to consult with the one in charge of that before merging. Otherwise, once the inventory is fully functioning with the PGH, I will merge to a dev and test it, and if it works, then I will merge into main